split cases: first round
