+++ src/business-types.ts
@@ -1,0 +1,12 @@
+export type UserId = string;
+type UserName = string;
+type EmailAddress = string;
+export type Profile = {
+  userId: UserId;
+  name: UserName;
+  emailAddress: EmailAddress;
+};
+export type EmailMessage = {
+  To: EmailAddress;
+  Body: string;
+};
